requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.3
deprecation==2.1.0
exceptiongroup==1.3.1
fastapi==0.123.8
flatbuffers==25.9.23
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-genai==1.53.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
ImageIO==2.37.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy_loader==0.4
llvmlite==0.45.1
mpmath==1.3.0
multidict==6.7.0
networkx==3.4.2
numba==0.62.1
numpy==2.2.6
onnxruntime==1.23.2
opencv-python-headless==4.12.0.88
packaging==25.0
pillow==12.0.0
platformdirs==4.5.0
pooch==1.8.2
postgrest==2.25.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
PyMatting==1.1.14
pyparsing==3.2.5
pyreadline3==3.5.4
python-dotenv==1.2.1
realtime==2.25.0
referencing==0.37.0
rembg==2.0.69
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
scikit-image==0.25.2
scipy==1.15.3
starlette==0.50.0
storage3==2.25.0
StrEnum==0.4.15
supabase==2.25.0
supabase-auth==2.25.0
supabase-functions==2.25.0
sympy==1.14.0
tenacity==9.1.2
tifffile==2025.5.10
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
yarl==1.22.0
